testing/requirements.txt
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
mysql-connector-python==8.2.0
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
PyJWT==2.8.0
